create NotificationToneManager to handle tone business logic
