+++ ElementX/Sources/Screens/Settings/NotificationSettingsScreen/View/NotificationSettingsScreen.swift
@@ -38,7 +38,9 @@ struct NotificationSettingsScreen: View {
                         additionalSettingsSection
                     }
 
-                    soundSelectionSection
+                    if context.viewState.canSelectTones {
+                        soundSelectionSection
+                    }
                 }
             }
         }
